fixed
fixed daniel return statements that were causing the cause to go into an infinite loop. A simple continue; was used to fix this issue. Had daniel taken the time to actually read through the code, he would have noticed that the return statement was incorrect. This shows that daniel did not take his part seriously.